Add files via upload
Fixes many of the formatting issues that were present in the first iteration of the UI. Adds in constraint for many objects so they remain in place while the app runs. Changed color scheme & font of the app design so it was continuous while remaining pleasing to the user's eye. 4/6 XML pages worth of progress.