Add files via upload
This file completes the last two XML pages in the UI (4/6 -> 6/6). This can be packed in with the backend & built into an APK for the first build of the app. Further, plans are to remove hardcoded strings and place them into their appropriate .xml page but it is not necessary to do for the first full test run.