Add files via upload
This has certain things that are more up to date than the current build, but certain things that are behind the current build (and is accordingly named Half Update). This version updates the positioning of certain buttons so that users see information correctly, and adds in animations for two sections of the home page.