Add files via upload
A minor update cleaning the app folders of extraneous files that have been subsumed by other, currently in use files. Also adjusts certain parts of two layouts for ease of use. Lastly changes the package name, versioncode/versionname, and build.Gradle file in preparation for the release of this app.